+++ app/src/main/java/sk/dejw/android/georiddles/utils/cursor/GameCursorUtils.java
@@ -21,7 +21,7 @@ public final class GameCursorUtils {
         ArrayList<Game> list = new ArrayList<>();
         while (cursor.moveToNext()) {
             Game game = new Game(
-                    cursor.getInt(GAME_ID),
+                    cursor.getLong(GAME_ID),
                     UUID.fromString(cursor.getString(GAME_UUID)),
                     cursor.getString(GAME_TITLE),
                     cursor.getString(GAME_CODE),
@@ -43,7 +43,7 @@ public final class GameCursorUtils {
         if (cursor.getCount() > 0) {
             cursor.moveToFirst();
             Game game = new Game(
-                    cursor.getInt(GAME_ID),
+                    cursor.getLong(GAME_ID),
                     UUID.fromString(cursor.getString(GAME_UUID)),
                     cursor.getString(GAME_TITLE),
                     cursor.getString(GAME_CODE),
